refactor(repos): split repos into subscription and weather repos
