+++ internal/repos/subscription/subscription_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/lib/pq"
 	"github.com/stretchr/testify/assert"
 	"github.com/velosypedno/genesis-weather-api/internal/domain"
-	"github.com/velosypedno/genesis-weather-api/internal/repos"
+	subr "github.com/velosypedno/genesis-weather-api/internal/repos/subscription"
 )
 
 const (
@@ -31,7 +31,7 @@ func TestCreateSubscription_Success(t *testing.T) {
 	assert.NoError(t, err)
 	defer closeDB(mock, db, t)
 
-	repo := repos.NewSubscriptionDBRepo(db)
+	repo := subr.NewSubscriptionDBRepo(db)
 
 	sub := domain.Subscription{
 		ID:        uuid.New(),
@@ -60,7 +60,7 @@ func TestCreateSubscription_EmailExists(t *testing.T) {
 	assert.NoError(t, err)
 	defer closeDB(mock, db, t)
 
-	repo := repos.NewSubscriptionDBRepo(db)
+	repo := subr.NewSubscriptionDBRepo(db)
 
 	sub := domain.Subscription{
 		ID:        uuid.New(),
@@ -92,7 +92,7 @@ func TestActivateSubscription_Success(t *testing.T) {
 	assert.NoError(t, err)
 	defer closeDB(mock, db, t)
 
-	repo := repos.NewSubscriptionDBRepo(db)
+	repo := subr.NewSubscriptionDBRepo(db)
 	token := uuid.New()
 
 	mock.ExpectExec(regexp.QuoteMeta(`UPDATE subscriptions SET activated = true WHERE token = $1`)).
@@ -111,7 +111,7 @@ func TestActivateSubscription_TokenNotFound(t *testing.T) {
 	}
 	defer closeDB(mock, db, t)
 
-	repo := repos.NewSubscriptionDBRepo(db)
+	repo := subr.NewSubscriptionDBRepo(db)
 	token := uuid.New()
 
 	mock.ExpectExec(regexp.QuoteMeta(`UPDATE subscriptions SET activated = true WHERE token = $1`)).
@@ -128,7 +128,7 @@ func TestDeleteSubscriptionByToken_Success(t *testing.T) {
 	assert.NoError(t, err)
 	defer closeDB(mock, db, t)
 
-	repo := repos.NewSubscriptionDBRepo(db)
+	repo := subr.NewSubscriptionDBRepo(db)
 	token := uuid.New()
 
 	mock.ExpectExec(regexp.QuoteMeta(`DELETE FROM subscriptions WHERE token = $1`)).
@@ -145,7 +145,7 @@ func TestDeleteSubscriptionByToken_TokenNotFound(t *testing.T) {
 	assert.NoError(t, err)
 	defer closeDB(mock, db, t)
 
-	repo := repos.NewSubscriptionDBRepo(db)
+	repo := subr.NewSubscriptionDBRepo(db)
 	token := uuid.New()
 
 	mock.ExpectExec(regexp.QuoteMeta(`DELETE FROM subscriptions WHERE token = $1`)).
@@ -162,7 +162,7 @@ func TestGetActivatedSubscriptionsByFreq_Success(t *testing.T) {
 	assert.NoError(t, err)
 	defer closeDB(mock, db, t)
 
-	repo := repos.NewSubscriptionDBRepo(db)
+	repo := subr.NewSubscriptionDBRepo(db)
 
 	freq := domain.FreqDaily
 
@@ -186,7 +186,7 @@ func TestGetActivatedSubscriptionsByFreq_QueryError(t *testing.T) {
 	assert.NoError(t, err)
 	defer closeDB(mock, db, t)
 
-	repo := repos.NewSubscriptionDBRepo(db)
+	repo := subr.NewSubscriptionDBRepo(db)
 
 	freq := domain.FreqDaily
