+++ internal/repos/weather/weather_test.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/velosypedno/genesis-weather-api/internal/domain"
-	"github.com/velosypedno/genesis-weather-api/internal/repos"
+	weathr "github.com/velosypedno/genesis-weather-api/internal/repos/weather"
 )
 
 type mockHTTPClient struct {
@@ -40,7 +40,7 @@ func TestGetCurrentWeather_Success(t *testing.T) {
 		},
 	}
 
-	repo := repos.NewWeatherAPIRepo("dummy-api-key", client)
+	repo := weathr.NewWeatherAPIRepo("dummy-api-key", client)
 	weather, err := repo.GetCurrent(context.Background(), "Kyiv")
 
 	assert.NoError(t, err)
@@ -66,7 +66,7 @@ func TestGetCurrentWeather_CityNotFound(t *testing.T) {
 		},
 	}
 
-	repo := repos.NewWeatherAPIRepo("dummy-api-key", client)
+	repo := weathr.NewWeatherAPIRepo("dummy-api-key", client)
 	_, err := repo.GetCurrent(context.Background(), "InvalidCity")
 
 	assert.ErrorIs(t, err, domain.ErrCityNotFound)
@@ -82,7 +82,7 @@ func TestGetCurrentWeather_APIKeyInvalid(t *testing.T) {
 		},
 	}
 
-	repo := repos.NewWeatherAPIRepo("invalid-api-key", client)
+	repo := weathr.NewWeatherAPIRepo("invalid-api-key", client)
 	_, err := repo.GetCurrent(context.Background(), "Kyiv")
 
 	assert.Error(t, err)
@@ -96,7 +96,7 @@ func TestGetCurrentWeather_HTTPError(t *testing.T) {
 		},
 	}
 
-	repo := repos.NewWeatherAPIRepo("dummy-api-key", client)
+	repo := weathr.NewWeatherAPIRepo("dummy-api-key", client)
 	_, err := repo.GetCurrent(context.Background(), "Kyiv")
 
 	assert.Error(t, err)
@@ -113,7 +113,7 @@ func TestGetCurrentWeather_BadJSON(t *testing.T) {
 		},
 	}
 
-	repo := repos.NewWeatherAPIRepo("dummy-api-key", client)
+	repo := weathr.NewWeatherAPIRepo("dummy-api-key", client)
 	_, err := repo.GetCurrent(context.Background(), "Kyiv")
 
 	assert.Error(t, err)
